+++ src/Infrastructure/Queue/IQueueProvider.cs
@@ -8,9 +8,9 @@ namespace ViajaNet.JobApplication.Infrastructure.Queue
     public interface IQueueProvider : IDisposable
     {
         void Push<TValue>(TValue data)
-            where TValue : class, new();
+            where TValue : class;
         void Push<TValue>(string topic, TValue data)
-            where TValue : class, new();
+            where TValue : class;
         void Shift(EventHandler<BasicDeliverEventArgs> handler);
         void Shift(string topic, EventHandler<BasicDeliverEventArgs> handler);
     }